+++ src/RcppExports.cpp
@@ -11,8 +11,8 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
 // detect_all_events
-RObject detect_all_events(DataFrame df, SEXP reading_minutes, bool sort_time, double inter_gap, bool return_interpolated);
-RcppExport SEXP _cgmguru_detect_all_events(SEXP dfSEXP, SEXP reading_minutesSEXP, SEXP sort_timeSEXP, SEXP inter_gapSEXP, SEXP return_interpolatedSEXP) {
+RObject detect_all_events(DataFrame df, SEXP reading_minutes, bool sort_time, double inter_gap, bool return_interpolated, std::string summary_metrics_source);
+RcppExport SEXP _cgmguru_detect_all_events(SEXP dfSEXP, SEXP reading_minutesSEXP, SEXP sort_timeSEXP, SEXP inter_gapSEXP, SEXP return_interpolatedSEXP, SEXP summary_metrics_sourceSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -21,7 +21,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< bool >::type sort_time(sort_timeSEXP);
     Rcpp::traits::input_parameter< double >::type inter_gap(inter_gapSEXP);
     Rcpp::traits::input_parameter< bool >::type return_interpolated(return_interpolatedSEXP);
-    rcpp_result_gen = Rcpp::wrap(detect_all_events(df, reading_minutes, sort_time, inter_gap, return_interpolated));
+    Rcpp::traits::input_parameter< std::string >::type summary_metrics_source(summary_metrics_sourceSEXP);
+    rcpp_result_gen = Rcpp::wrap(detect_all_events(df, reading_minutes, sort_time, inter_gap, return_interpolated, summary_metrics_source));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -269,7 +270,7 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_cgmguru_detect_all_events", (DL_FUNC) &_cgmguru_detect_all_events, 5},
+    {"_cgmguru_detect_all_events", (DL_FUNC) &_cgmguru_detect_all_events, 6},
     {"_cgmguru_detect_between_maxima", (DL_FUNC) &_cgmguru_detect_between_maxima, 2},
     {"_cgmguru_detect_hyperglycemic_events", (DL_FUNC) &_cgmguru_detect_hyperglycemic_events, 10},
     {"_cgmguru_detect_hypoglycemic_events", (DL_FUNC) &_cgmguru_detect_hypoglycemic_events, 9},